Update index.php
i have changed the spacing between buttons the search bar then added a dropdown filter location for the job listing, made a dummy job card with an expandable toggle function by either pressing the title or by the toggle sign on the right which opens a job description with the option to apply now which hasnt yet been implemented, i added a employers button section in the navigation as when someone wants to add a new job it would be from there, also i changed the title from find your dream job to find a job as it is for side projects and wouldnt been a dream job i kept the php html css and javascript in one file to make it easier. i kept the design to what was originally inputed. also when you input a job title that isnt available a message will be left behind. Also alert function have been made for the buttons that have had no implementation yet.