Updated Payment page
There is now a working option to pay with both PayPal and card via PayPal. There is also an option to pay bank through another provider which the client can choose to keep or remove! The PayPal link and stripe link will have to be removed and replaced with clients chosen address which is currently not available however, PayPal is in sandbox mode which I found a link to online, this allows us to test that PayPal works which it does however will need to be changed to the clients address,